create a new download_manager.prompts module
For better readability.
